Another batch of improvements after pointing dechk at our own headers

+++ headers/stdlib.h
@@ -39,8 +39,8 @@ ldiv_t;
 
 typedef struct
 {
-  long quot;
-  long rem;
+  long long quot;
+  long long rem;
 }
 lldiv_t;
